refactor(#35) 오타 수정
